perf(mypyc): compile prometheus observer
