+++ tests/unit/utils/test_mypyc_inventory.py
@@ -124,6 +124,7 @@ def test_inventory_records_rest_of_mypyc_boundary_decisions() -> None:
     assert "sqlspec/storage/_paths.py" in payload["compiled_modules"]
     assert "sqlspec/storage/_utils.py" in payload["compiled_modules"]
     assert "sqlspec/base.py" in payload["compiled_modules"]
+    assert "sqlspec/extensions/prometheus/_observer.py" in payload["compiled_modules"]
     assert "sqlspec/data_dictionary/_loader.py" in payload["compiled_modules"]
     assert "sqlspec/data_dictionary/dialects/bigquery.py" in payload["compiled_modules"]
     assert "sqlspec/data_dictionary/dialects/cockroachdb.py" in payload["compiled_modules"]
@@ -184,7 +185,6 @@ def test_inventory_records_wave4_candidate_and_hard_block_buckets() -> None:
     expected_candidates = {
         "sqlspec/extensions/fastapi/providers.py",
         "sqlspec/extensions/litestar/providers.py",
-        "sqlspec/extensions/prometheus/__init__.py",
         "sqlspec/storage/backends/fsspec.py",
         "sqlspec/storage/backends/local.py",
         "sqlspec/storage/backends/obstore.py",
@@ -208,6 +208,12 @@ def test_inventory_records_wave4_candidate_and_hard_block_buckets() -> None:
     assert hot_surfaces["sqlspec/base.py"]["surface"] == "compiled"
     assert hot_surfaces["sqlspec/base.py"]["status"] == "compiled"
     assert hot_surfaces["sqlspec/base.py"]["reason"]
+    assert hot_surfaces["sqlspec/extensions/prometheus/_observer.py"]["surface"] == "compiled"
+    assert hot_surfaces["sqlspec/extensions/prometheus/_observer.py"]["status"] == "compiled"
+    assert hot_surfaces["sqlspec/extensions/prometheus/_observer.py"]["reason"]
+    assert hot_surfaces["sqlspec/extensions/prometheus/__init__.py"]["surface"] == "keep_interpreted"
+    assert hot_surfaces["sqlspec/extensions/prometheus/__init__.py"]["status"] == "interpreted"
+    assert hot_surfaces["sqlspec/extensions/prometheus/__init__.py"]["reason"]
 
     for module_path in expected_hard_blocks:
         details = hot_surfaces[module_path]
